fix: Most corrupt chunks gone, move away from NIO
- Sometimes invisible chunks still appear, trying to figure that out.

+++ components/gfx-opengl/src/main/java/net/caffeinemc/gfx/opengl/device/GlRenderDevice.java
@@ -309,7 +309,7 @@ public class GlRenderDevice implements RenderDevice {
         }
 
         @Override
-        public void multiDrawElementsIndirect(PrimitiveType primitiveType, ElementFormat elementType, int indirectOffset, int indirectCount) {
+        public void multiDrawElementsIndirect(PrimitiveType primitiveType, ElementFormat elementType, long indirectOffset, int indirectCount) {
             if (RenderConfiguration.API_CHECKS) {
                 Validate.notNull(this.elementBuffer, "Element buffer target not bound");
                 Validate.notNull(this.commandBuffer, "Command buffer target not bound");
